+++ src/vue/pages/polls-all/PollVoteForm.vue
@@ -234,6 +234,7 @@ export default {
 .poll-vote-form__question {
   line-height: 1.4865;
   font-size: 1.6rem;
+  color: $col-text;
 }
 
 .poll-vote-form__choices {