+++ src/vue/navigation/Footer.vue
@@ -63,7 +63,11 @@ export default {
 }
 
 .footer__link {
-  color: $col-link;
+  color: $col-footer-text;
+}
+
+.footer__text {
+  color: $col-footer-text;
 }
 
 .footer__language-picker,