+++ CHANGELOG.md
@@ -11,6 +11,9 @@ for further information about branching and tagging conventions.
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 - A bug with infinite loading of the page "Register of shares"
 
+#### Changed
+- Color scheme improvements
+
 ## [1.10.3-rc.0] - 2019-10-11
 #### Added
 - Integration with Erc20 and stellar