+++ CHANGELOG.md
@@ -7,13 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
+
 #### Fixed
 - A bug when the user can not update asset with deposit withdraw integration
+
+#### Changed
+- Moved logo and favicon to branding folder
+
+### "Under the hood" changes
+#### Added
+- `APP_NAME` in config
 
 ## [1.10.3-rc.1] - 2019-10-18
 #### Fixed
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 - A bug with infinite loading of the page "Register of shares"
+
+#### Changed
+- Color scheme improvements
 
 ## [1.10.3-rc.0] - 2019-10-11
 #### Added