+++ CHANGELOG.md
@@ -7,12 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
-### "Under the hood" changes
-#### Added
-- `APP_NAME` in config
+
+#### Fixed
+- A bug when the user can not update asset with deposit withdraw integration
 
 #### Changed
 - Moved logo and favicon to branding folder
+
+### "Under the hood" changes
+#### Added
+- `APP_NAME` in config
 
 ## [1.10.3-rc.1] - 2019-10-18
 #### Fixed